refactored and added minor updates to user auth
