+++ src/Pages/logIn.cshtml.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using ePiggyWeb.Authentication;
+using ePiggyWeb.DataBase;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Mvc;
@@ -31,7 +31,7 @@ namespace ePiggyWeb.Pages
 
         public async Task<IActionResult> OnPost(string returnUrl)
         {
-            var id = UserAuth.Login(Email, Password);
+            var id = UserDatabase.Login(Email, Password);
             if (id > -1)
             {
                 var claims = new List<Claim>
